feat: add component sets joints on cell edit
